[@layer] User layer order to resolve @font-face family name conflicts
@font-face rules and other name-defining at rules are different that,
rules with the same font-family name do not completely override each
other, but have different priorities in font fallback.

Therefore, this patch modifies CSSSegmentedFontFace to store CSS-
connected font faces in the cascade layer order, so that font faces in
a higher layer will be attempted first in font fallback.

Bug: 1096765
Change-Id: Iccd84b0fd48f1e13eb536a48eabc7f22ccd32149