Add Echo Stream WPTs for WebTransport
This CL adds 3 test cases:
- echo data on client-initiated bidirectional streams
- echo data on server-initiated bidirectional streams
- echo data using client and server-initiated unidirectional streams

The echo server handler has also been modified to add support for
creating server-initiated bidirectional streams when a WebTransport
session is established. Furthermore, when stream data is received
from a client-initiated unidirectional stream, the handler now keeps
a record of incoming and outgoing unidirectional streams, and creates
a new unidirectional stream when needed to echo back data to the
client. The unidirectional streams are mapped using the session's id
and incoming stream id as keys.

Bug: 1201569
Change-Id: Iaa7d412d090be96c2ad42da2cad416e5cc60a659