Don't look for a sibling of a resumed node if we've seen all children.
NGBlockChildIterator would call NGBlockNode::NextSibling() in an inline
formatting context when we were at the last child break token for a
float. Floats become regular block children when resumed, so we cannot
really call NextSibling() on them at all. And we don't have to, since
there'll either be a break token to tell us where to resume in the
inline formatting context, if there is anything at all to resume.
Otherwise HasSeenAllChildren() will return true, and we can stop.

Fix and clean up NGBlockChildIterator, to hopefully make it easier to
follow.

Bug: 1249828
Change-Id: I91d03a2bfd0394a594d99ba512a992f29b4848ba